+++ src/app/client/client-list.tsx
@@ -4,22 +4,23 @@ import { Client } from "./client";
 
 interface ClientListProps {
   clients: Client[];
-  setClient(clientId: string): void;
 }
 
-const ClientList = ({ clients, setClient }: ClientListProps) => (
-  <ul>
-    {clients.map((client, index) => (
-      <li key={index}>
-        <button
-          className="btn btn-blue my-3"
-          onClick={() => setClient(client.clientId)}
-        >
-          {client.name}
-        </button>
-      </li>
-    ))}
-  </ul>
+const ClientList = ({ clients }: ClientListProps) => (
+  <table>
+    <thead>
+      <th>Client Id</th>
+      <th>Client Name</th>
+    </thead>
+    <tbody>
+      {clients.map((client, index) => (
+        <tr key={index}>
+          <td>{client.clientId}</td>
+          <td>{client.name}</td>
+        </tr>
+      ))}
+    </tbody>
+  </table>
 );
 
 export { ClientList };